[hail] Refactor LowerTableIR to use new LoweringPipeline infrastructure.
Delete local backend.
